[#28] rename the timestamp counter to accumulator
Up to the version 3.1.3, the DefaultTimestampStrategy used a counter to
fill the lower bits of the timestamp. The counter was incremented by ONE
when the method was invoked within the same millisecond. Now this job is
done by an accumulator that is incremented in each repetition with
different values.

Whenever the method is invokated within the same millisecond, the
ellapsed ticks are added to the accumulator, i.e., the count of
100-nsecs intervals since the last call. The method's returning
timestamp is the the current milliseconds plus the accumulator value.

TIMESTAMP = MILLISECONDS * 10000 + ACCUMULATOR;