[#28] rename the timestamp counter to accumulator
Until version 3.1.3, DefaultTimestampStrategy used a counter to fill the
lower bits of the timestamp. The counter was incremented by 1 when the
method was invoked in the same millisecond. Now this work is done by an
accumulator that is increased at each repetition with different values.

Whenever the method is invoked in the same millisecond, the elapsed
ticks are added to the accumulator, that is, the 100-nsec interval count
since the last call. The method's return is the current milliseconds
plus the accumulator value.

TIMESTAMP = MILLISECONDS * 10000 + ACCUMULATOR;