feat: UUID Comparators (feature request) #62
Added `UuidComparator` for comparing and sorting UUIDs.

It has two static methods:

1. The default static method compares two time-based UUIDs by comparing
the timestamps first and then comparing the least significant bits as
unsigned 64-bit integers. If both UUIDs are not time-based then it
compares them as unsigned 128-bit integers.

2. The opaque static method compares two UUIDs as unsigned 128-bit
integers. It's the same as lexicographic sorting of UUID canonical
strings.