perf: BaseN encoding speed improvement #63
Added `CustomDivisor` to be used with `BaseNCodec`.

It is a division function interface that returns quotient and remainder.

It can be used to boost encoding speed utilizing specialized libraries
like `libdivide`, a library for optimizing integer division.

Adding `CustomDivisor` did not affect performance. In fact, the
refactoring done to include it improved performance by 15% for encoding
and 8% for decoding.

Simple usage example:

```
CustomDivider divideBy64 = x -> new long[] { x / 64, x % 64 };
```