Add a unique number to be XORed with the thread-local random sequence
The JVM unique number tries to mitigate the fact that the thread
local random is not seeded with a secure random seed by default.
Their seeds are based on temporal data and predefined constants.
Although the seeds are unique per JVM, they are not across JVMs.
It helps to generate different sequences of numbers even if two
ThreadLocalRandom are by chance instantiated with the same seed.
Of course it doesn't better the output, but doesn't hurt either.